Pulled out middleware present check into its own function.
And fixed Python 3 compat.
