+++ mezzanine/utils/cache.py
@@ -1,18 +1,15 @@
 from __future__ import unicode_literals
 
-import types
 from hashlib import md5
-from inspect import getmro
 from time import time
 
 from django.core.cache import cache
 from django.utils.lru_cache import lru_cache
 from django.utils.cache import _i18n_cache_key_suffix
-from django.utils.module_loading import import_string
 
 from mezzanine.conf import settings
-from mezzanine.utils.deprecation import get_middleware_setting
 from mezzanine.utils.sites import current_site_id
+from mezzanine.utils.conf import middlewares_or_subclasses_installed
 
 
 def _hashed_key(key):
@@ -68,22 +65,11 @@ def cache_installed():
     """
     has_key = bool(getattr(settings, "NEVERCACHE_KEY", ""))
 
-    def flatten(seqs):
-        return (item for seq in seqs for item in seq)
-
-    middleware_items = map(import_string, get_middleware_setting())
-    middleware_classes = filter(
-        lambda m: isinstance(m, (types.ClassType, types.TypeType)),
-        middleware_items)
-    middleware_ancestors = set(flatten(map(getmro, middleware_classes)))
-
-    mezzanine_cache_middleware_classes = {
-        import_string("mezzanine.core.middleware.UpdateCacheMiddleware"),
-        import_string("mezzanine.core.middleware.FetchFromCacheMiddleware"),
-    }
-
     return (has_key and settings.CACHES and not settings.TESTING and
-            mezzanine_cache_middleware_classes.issubset(middleware_ancestors))
+            middlewares_or_subclasses_installed([
+                "mezzanine.core.middleware.UpdateCacheMiddleware",
+                "mezzanine.core.middleware.FetchFromCacheMiddleware",
+            ]))
 
 
 def cache_key_prefix(request):
